+++ R/t.error.R
@@ -1,4 +1,8 @@
 t.error <-
 function(parameters, values, probabilities, weights, degreesfreedom){
-	sum(weights * (pt((values-parameters[1]) / exp(parameters[2]), degreesfreedom) - probabilities)^2)
+	sum(weights * (pt((values-parameters[1]) / exp(parameters[2]),
+	                  degreesfreedom) - probabilities)^2)
+
 }
+#   sum(weights * ((qt(probabilities, degreesfreedom) * exp(parameters[2]) +
+# parameters[1]) - values)^2)